+++ src/env.ts
@@ -32,6 +32,9 @@ export const CELATONE_API_OVERRIDE =
 export const HASURA_ADMIN_SECRET =
   process.env.NEXT_PUBLIC_HASURA_ADMIN_SECRET ?? "";
 
+export const CELATONE_VERIFICATION_API =
+  process.env.NEXT_PUBLIC_CELATONE_VERIFICATION_API ?? "";
+
 export const INITIA_MOVE_DECODER =
   process.env.NEXT_PUBLIC_INITIA_MOVE_DECODER ?? "";