+++ src/lib/app-fns/tx/storeCode.tsx
@@ -6,13 +6,13 @@ import type { Observable } from "rxjs";
 import { EstimatedFeeRender } from "lib/components/EstimatedFeeRender";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
-import { TxStreamPhase } from "lib/types";
 import type {
   BechAddr,
   ComposedMsg,
   Option,
   TxResultRendering,
 } from "lib/types";
+import { TxStreamPhase } from "lib/types";
 import { feeFromStr, findAttr } from "lib/utils";
 
 import { catchTxError } from "./common";