+++ src/lib/components/modal/index.ts
@@ -6,3 +6,5 @@ export * from "./ActionModal";
 export * from "./NoMobile";
 export * from "./RedoModal";
 export * from "./UnsupportedTokensModal";
+export * from "./wasm-verify-status";
+export * from "./wasm-verify-submit";